Tweak the svn log --xml output to be more valid.
Suggested by: Lieven Govaerts <lgo@mobsol.be>

This change makes the author and date elements in logentry optional, which they
should have been since the beginning.  While here, we also make the DTD
slightly stricter regarding the path element in paths elements to avoid
confusion.  We also ensure that an empty date element is not output even if
the date string is empty, which the API allows.

NOTE: This change makes the DTD looser, which might break scripts, but I
consider the previous content model of logentry to be buggy.

* subversion/clients/cmdline/log-cmd.c (log_message_receiver_xml): Avoid
  outputting an empty date element.

* subversion/clients/cmdline/dtd/log.dtd (logentry): Make author and
  date optional.
  (paths): Require at least one path element (the paths element itself
  is optional and it makes no sense to have two ways to express the same
  thing).

* subversion/tests/clients/cmdline/log_tests.py
  (log_xml_empty_date): New test.
  (test_list): Run it.

* subversion/tests/clients/cmdline/prop_tests.py
  (revprop_change): Factor out some code from here...
* subversion/tests/clients/cmdline/svntest/actions.py
  (enable_revprop_changes): ... to here.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/clients/cmdline@855705 13f79535-47bb-0310-9956-ffa450edef68