Rewrite Skip to wrap a TestCase instead of inheriting from _Predicate.
Previously, Skip inherited from _Predicate, and due to the peculiar
implementation of _Predicate, this only allowed Skip to wrap
_Predicates or naked functions, but not arbitrary TestCases.

This commit changes Skip to inherit directly from TestCase and to wrap
an arbitrary instance of TestCase.  Most methods are delegated to the
wrapped TestCase, and some are implemented in Skip to modify the
behavior.

* subversion/tests/cmdline/svntest/testcase.py (Skip): Rewritten as
  described above.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@859351 13f79535-47bb-0310-9956-ffa450edef68