Before running each command-line test, save the current directory, and restore it after the test has finished.
Given that, it's unnecessary to wrap every test that changes the
directory in a try/finally block to change it back, so remove all the
relevant try/finallys from the command-line test suite.  When this
leaves a call to os.chdir as the last line of a test, remove the
os.chdir as well; if os.getcwd is called just to obtain an argument
for a removed os.chdir call, remove the os.getcwd call as well.

[in subversion/tests/cmdline]

* svntest/main.py
  Remove obsolete comment implying that os was only imported for
   popen2.
  (TestRunner.run): Save current directory before invoking test and
   restore it afterwards.

* revert_tests.py, svnadmin_tests.py, log_tests.py, switch_tests.py,
  externals_tests.py, diff_tests.py, copy_tests.py, stat_tests.py,
  blame_tests.py, basic_tests.py, update_tests.py, commit_tests.py,
  schedule_tests.py, merge_tests.py, depth_tests.py:
  Remove try/finally from many tests, as described above.

Suggested by: malcolm
              dlr


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@865429 13f79535-47bb-0310-9956-ffa450edef68