Merge the 'mergeinfoless-copies' branch into trunk. (Merged /branches/mergeinfoless-copies:r27771-28001.)
This branch changes the way copies and merges happen in a fairly
significant want.  

First, copies revert to their pre-1.5 state of simply being copies,
with the mild exception that some no-cost-to-calculate mergeinfo is
dropped into place on the copy destination.  (As a side-effect, we
lose the --use-merge-history/-g option for 'svn copy' and 'svn move'
as unnecessary.)

Secondly, merges now use node ancestry in their calculations, treating
a node's "mergeinfo" as the sum of its recorded mergeinfo (either
recorded directly or inherited from some parent directory) and its
natural history (the changes made directly to this node over time).

The result is a tradeoff: the restoration of copy's incredible speed
(which is a major selling point of Subversion) for a
slower-than-already-slow merge.  But the benefits of this change also
include theoretical correctness, algorithmic consistency, and -- best
of all -- that repositories created prior to Subversion 1.5 will
already have valuable information to feed into the merge algorithms of
Subversion 1.5 without having ever recorded the first svn:mergeinfo
property -- the natural history of branches!


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@868076 13f79535-47bb-0310-9956-ffa450edef68