Merge the tc-issue-3334 branch to trunk:
Fix part of issue #3334 "Tree conflict merry-go-round on update/switch".

This is the first case to be implemented: When "update" or "switch" deletes
something that is locally edited, as well as raising a tree conflict, the local
node should be scheduled for re-addition as a copy of the previously existing
version, so the user can keep their version simply by issuing an "svn resolved"
command.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@874774 13f79535-47bb-0310-9956-ffa450edef68