Read svn properties in XML rather than text format in test suite.
This makes the routine more robust to strange property values.  It also
makes it immune to changes in the text output format, as happened
between svn 1.5 and 1.6.  (This is useful for the cvs2svn project,
which uses the svntest infrastructure but doesn't want to care what
version of svn the user has installed.)

* subversion/tests/cmdline/svntest/tree.py
  (get_props): Read svn properties via "svn proplist -v" using the
  "--xml" option.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@878560 13f79535-47bb-0310-9956-ffa450edef68