Tweak “svnrdump load” tests to pass with svnadmin 1.6
svnadmin 1.6 does not dump properties in alphabetical order, but
svnadmin 1.7 and “svnrdump dump” do.

In this 1.6-backporting branch, let’s tolerate reordered lines
in the “svnadmin dump” output (though not the “svnradmin dump” output,
of course).  After that change, all tests expected to pass do pass.

Theoretically the resulting repository could be broken in some
way (e.g., reordered commits), so it would be nice to check that
“svnrdump load” followed by “svnrdump dump” is still the identity;
but this patch does not do that yet.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>