svn17_compat: Make use of the svn_ctype_is* functions more uniform
Merge r985014, r985669, and r987893 from the performance branch, which make
use of the svn_ctype_is* functions more uniform.

This also adds an include to subversion/libsvn_fs_fs/fs_fs.c which was added
as part of r983430 on the performance branch.  That revision is not related
to the above, and so isn't being merged here.

From subversion/trunk/subversion/libsvn_subr/dirent_uri.c@995212

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>