Merge the atomic-revprop branch to trunk.
This branch adds the svn_ra_change_rev_prop2() API, which enables revision
property edits to be atomic: the caller provides an old value and the FS layer
will only allow the edit to go through if the value it sees is that old value.
(The FS layer support, svn_fs_change_rev_prop2(), has been in trunk since
before the branch was started.)

It then uses the new API in svnsync to set the svn:sync-lock mutex on the
target repository, addressing issue #3546.


For further details, see:

* the BRANCH-README [1];
* the issue [2];
* periodic dev@ threads during the lifetime of the branch;
* the log messages on the branch.

[1] http://svn.apache.org/repos/asf/subversion/branches/atomic-revprop/BRANCH-README?p=1000359
[2] http://subversion.tigris.org/issues/show_bug.cgi?id=3546

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk/subversion/tests/cmdline@1000693 13f79535-47bb-0310-9956-ffa450edef68